+++ Sources/Extensions/UIKit/UIViewExtensions.swift
@@ -555,5 +555,18 @@ public extension UIView {
 		anchorCenterYToSuperview()
 	}
 
+    /// SwifterSwift: Search all superviews until a view with this class is found.
+    ///
+    /// - Parameter name: class of the view to search.
+    public func ancestorView<T: UIView>(withClass name: T.Type) -> T? {
+        guard let superview = superview else {
+            return nil
+        }
+        guard let view = superview as? T else {
+            return superview.ancestorView(withClass: name)
+        }
+        return view
+    }
+
 }
 #endif